+++ clevercsv/console/commands/code.py
@@ -41,10 +41,12 @@ and copy the generated code to a Python script.
             encoding=encoding,
             verbose=self.option("verbose"),
         )
+        if dialect is None:
+            return self.line("Dialect detection failed.")
 
         d = f'"{dialect.delimiter}"'
         q = '"%s"' % (dialect.quotechar.replace('"', '\\"'))
-        e = f'"{dialect.escapechar}"'
+        e = repr(f"{dialect.escapechar}").replace("'", '"')
         base = [
             "",
             f"# Code generated with CleverCSV version {__version__}",