+++ clevercsv/console/commands/view.py
@@ -4,6 +4,7 @@ import tabview
 
 from cleo import Command
 
+from clevercsv.exceptions import NoDetectionResult
 from clevercsv.wrappers import read_csv
 
 from ._utils import parse_int
@@ -28,12 +29,15 @@ Use the <info>view</info> command to view a CSV file on the command line.
     def handle(self):
         verbose = self.io.verbosity > 0
         num_chars = parse_int(self.option("num-chars"), "num-chars")
-        rows = read_csv(
-            self.argument("path"),
-            encoding=self.option("encoding"),
-            num_chars=num_chars,
-            verbose=verbose,
-        )
-        if self.option('transpose'):
+        try:
+            rows = read_csv(
+                self.argument("path"),
+                encoding=self.option("encoding"),
+                num_chars=num_chars,
+                verbose=verbose,
+            )
+        except NoDetectionResult:
+            self.line("Dialect detection failed.")
+        if self.option("transpose"):
             rows = list(map(list, zip(*rows)))
         tabview.view(rows)