+++ clevercsv/exceptions.py
@@ -9,5 +9,10 @@ Author: Gertjan van den Burg
 
 from .cparser import Error as ParserError
 
+
 class Error(ParserError):
+    pass
+
+
+class NoDetectionResult(Exception):
     pass