+++ tests/test_unit/test_cparser.py
@@ -8,7 +8,6 @@ Author: Gertjan van den Burg
 """
 
 import io
-import six
 import unittest
 
 from clevercsv.cparser_util import parse_data
@@ -21,10 +20,7 @@ class ParserTestCase(unittest.TestCase):
     """
 
     def _parse_test(self, string, expect, **kwargs):
-        if six.PY2:
-            buf = io.StringIO(string.decode("ascii"), newline="")
-        else:
-            buf = io.StringIO(string, newline="")
+        buf = io.StringIO(string, newline="")
         result = list(parse_data(buf, **kwargs))
         self.assertEqual(result, expect)