+++ tests/test_unit/test_potential_dialects.py
@@ -10,10 +10,10 @@ Author: Gertjan van den Burg
 import unittest
 
 from clevercsv.potential_dialects import (
-    masked_by_quotechar,
     filter_urls,
     get_delimiters,
     get_quotechars,
+    masked_by_quotechar,
 )
 
 
@@ -41,5 +41,6 @@ class PotentialDialectTestCase(unittest.TestCase):
         out = get_delimiters(data, "UTF-8")
         self.assertEqual(out, exp)
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     unittest.main()