+++ tests/test_unit/test_write.py
@@ -9,23 +9,22 @@ Author: Gertjan van den Burg
 
 
 import clevercsv
+import tempfile
 import unittest
 
 from clevercsv.dialect import SimpleDialect
 
-from tempfile import TemporaryFile
-
 
 class WriterTestCase(unittest.TestCase):
     def writerAssertEqual(self, input, expected_result):
-        with TemporaryFile("w+", newline="") as fileobj:
+        with tempfile.TemporaryFile("w+", newline="") as fileobj:
             writer = clevercsv.writer(fileobj, dialect=self.dialect)
             writer.writerows(input)
             fileobj.seek(0)
             self.assertEqual(fileobj.read(), expected_result)
 
     def _write_test(self, fields, expect, **kwargs):
-        with TemporaryFile("w+", newline="") as fileobj:
+        with tempfile.TemporaryFile("w+", newline="") as fileobj:
             writer = clevercsv.writer(fileobj, **kwargs)
             writer.writerow(fields)
             fileobj.seek(0)
@@ -34,7 +33,7 @@ class WriterTestCase(unittest.TestCase):
             )
 
     def _write_error_test(self, exc, fields, **kwargs):
-        with TemporaryFile("w+", newline="") as fileobj:
+        with tempfile.TemporaryFile("w+", newline="") as fileobj:
             writer = clevercsv.writer(fileobj, **kwargs)
             with self.assertRaises(exc):
                 writer.writerow(fields)
@@ -45,7 +44,9 @@ class WriterTestCase(unittest.TestCase):
         self._write_error_test(clevercsv.Error, None)
         self._write_test((), "")
         self._write_test([None], '""')
-        self._write_error_test(clevercsv.Error, [None], quoting=clevercsv.QUOTE_NONE)
+        self._write_error_test(
+            clevercsv.Error, [None], quoting=clevercsv.QUOTE_NONE
+        )
         # Check that exceptions are passed up the chain
         class BadList:
             def __len__(self):
@@ -83,7 +84,9 @@ class WriterTestCase(unittest.TestCase):
         self._write_test(
             ["a", 1, "p,q"], '"a","1","p,q"', quoting=clevercsv.QUOTE_ALL
         )
-        self._write_test(["a\nb", 1], '"a\nb","1"', quoting=clevercsv.QUOTE_ALL)
+        self._write_test(
+            ["a\nb", 1], '"a\nb","1"', quoting=clevercsv.QUOTE_ALL
+        )
 
     def test_write_simpledialect(self):
         self._write_test(
@@ -92,5 +95,6 @@ class WriterTestCase(unittest.TestCase):
             dialect=SimpleDialect(delimiter=",", quotechar="|", escapechar=""),
         )
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     unittest.main()