+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## Version 0.4.0
+
+* Update URL regular expression and add unit tests
+* Add IPv4 type detection
+* Add tie-breaker for combined quotechar and escapechar ties
+
+## Version 0.3.7
+
+* Bugfix for console script ``code`` command
+* Update readme
+
+## Version 0.3.6
+
+* Cleanly handle failure to detect dialect in console application
+* Remove any (partial) support for Python 2
+
 ## Version 0.3.5
 
 * Remove Python parser - this speeds up file reading and tie breaking